common entities factory, final managers and systems updated to use bitmasks

+++ src/Entities/Components/UI/UIButtonComponent.h
@@ -9,7 +9,7 @@ struct UIButtonComponent
 {
     bool isSelected = false;
 
-    uint16_t defaultColor;
-    uint16_t selectedColor;
+    uint8_t defaultColor;
+    uint8_t selectedColor;
     std::function<void(GameManager& manager)> OnClicked; //might be empty and just depach events
 };
